backend/requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
